# packages/api-contracts/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MotorCove read API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/sales": {
      "get": {
        "responses": {
          "200": {
            "description": "Projected sales",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalesEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Read model unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sales/{saleId}": {
      "get": {
        "parameters": [
          {
            "name": "saleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^(0|[1-9]\\d*)$",
              "maxLength": 78,
              "description": "Canonical decimal uint256"
            }
          },
          {
            "name": "deploymentId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^0x[0-9a-fA-F]{64}$"
            }
          },
          {
            "name": "observeTxHash",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^0x[0-9a-fA-F]{64}$"
            }
          },
          {
            "name": "observeBlockNumber",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(0|[1-9]\\d*)$"
            }
          },
          {
            "name": "observeBlockHash",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^0x[0-9a-fA-F]{64}$"
            }
          },
          {
            "name": "observeLogIndex",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Projected sale or selector-scoped funding observation",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/SaleEnvelope"
                    },
                    {
                      "$ref": "#/components/schemas/FundingObservationEnvelope"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Sale not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Deployment mismatch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Read model unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SaleEnvelope": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "saleId": {
                "type": "string",
                "pattern": "^(0|[1-9]\\d*)$"
              },
              "tokenId": {
                "type": "string",
                "pattern": "^(0|[1-9]\\d*)$"
              },
              "seller": {
                "type": "string",
                "pattern": "^0x[0-9a-fA-F]{40}$"
              },
              "buyer": {
                "anyOf": [
                  {
                    "type": "string",
                    "pattern": "^0x[0-9a-fA-F]{40}$"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "priceWei": {
                "type": "string",
                "pattern": "^(0|[1-9]\\d*)$"
              },
              "fundedAt": {
                "anyOf": [
                  {
                    "type": "string",
                    "pattern": "^(0|[1-9]\\d*)$"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "expiresAt": {
                "anyOf": [
                  {
                    "type": "string",
                    "pattern": "^(0|[1-9]\\d*)$"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "status": {
                "type": "string",
                "enum": [
                  "LISTED",
                  "FUNDED",
                  "COMPLETED",
                  "CANCELLED",
                  "EXPIRED"
                ]
              },
              "tokenReclaimed": {
                "type": "boolean"
              },
              "metadataStatus": {
                "type": "string",
                "enum": [
                  "AVAILABLE",
                  "MISSING"
                ]
              },
              "catalogId": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "claim": {
                "anyOf": [
                  {
                    "type": "object",
                    "properties": {
                      "kind": {
                        "type": "string",
                        "enum": [
                          "SELLER_PROCEEDS",
                          "BUYER_REFUND"
                        ]
                      },
                      "status": {
                        "type": "string",
                        "enum": [
                          "NONE",
                          "CLAIMABLE",
                          "WITHDRAWN"
                        ]
                      },
                      "beneficiary": {
                        "type": "string",
                        "pattern": "^0x[0-9a-fA-F]{40}$"
                      },
                      "amountWei": {
                        "type": "string",
                        "pattern": "^(0|[1-9]\\d*)$"
                      }
                    },
                    "required": [
                      "kind",
                      "status",
                      "beneficiary",
                      "amountWei"
                    ],
                    "additionalProperties": false
                  },
                  {
                    "type": "null"
                  }
                ]
              }
            },
            "required": [
              "saleId",
              "tokenId",
              "seller",
              "buyer",
              "priceWei",
              "fundedAt",
              "expiresAt",
              "status",
              "tokenReclaimed",
              "metadataStatus",
              "catalogId",
              "claim"
            ],
            "additionalProperties": false
          },
          "provenance": {
            "type": "object",
            "properties": {
              "deploymentId": {
                "type": "string",
                "pattern": "^0x[0-9a-fA-F]{64}$"
              },
              "indexedBlockNumber": {
                "type": "string",
                "pattern": "^(0|[1-9]\\d*)$"
              },
              "indexedBlockHash": {
                "type": "string",
                "pattern": "^0x[0-9a-fA-F]{64}$"
              },
              "projectorVersion": {
                "type": "string",
                "pattern": "^(0|[1-9]\\d*)$"
              },
              "projectionBuildId": {
                "type": "string",
                "minLength": 1
              },
              "logScopeHash": {
                "type": "string",
                "pattern": "^0x[0-9a-fA-F]{64}$"
              }
            },
            "required": [
              "deploymentId",
              "indexedBlockNumber",
              "indexedBlockHash",
              "projectorVersion",
              "projectionBuildId",
              "logScopeHash"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "data",
          "provenance"
        ],
        "additionalProperties": false
      },
      "FundingObservationEnvelope": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "sale": {
                "anyOf": [
                  {
                    "type": "object",
                    "properties": {
                      "saleId": {
                        "type": "string",
                        "pattern": "^(0|[1-9]\\d*)$"
                      },
                      "tokenId": {
                        "type": "string",
                        "pattern": "^(0|[1-9]\\d*)$"
                      },
                      "seller": {
                        "type": "string",
                        "pattern": "^0x[0-9a-fA-F]{40}$"
                      },
                      "buyer": {
                        "anyOf": [
                          {
                            "type": "string",
                            "pattern": "^0x[0-9a-fA-F]{40}$"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "priceWei": {
                        "type": "string",
                        "pattern": "^(0|[1-9]\\d*)$"
                      },
                      "fundedAt": {
                        "anyOf": [
                          {
                            "type": "string",
                            "pattern": "^(0|[1-9]\\d*)$"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "expiresAt": {
                        "anyOf": [
                          {
                            "type": "string",
                            "pattern": "^(0|[1-9]\\d*)$"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "status": {
                        "type": "string",
                        "enum": [
                          "LISTED",
                          "FUNDED",
                          "COMPLETED",
                          "CANCELLED",
                          "EXPIRED"
                        ]
                      },
                      "tokenReclaimed": {
                        "type": "boolean"
                      },
                      "metadataStatus": {
                        "type": "string",
                        "enum": [
                          "AVAILABLE",
                          "MISSING"
                        ]
                      },
                      "catalogId": {
                        "type": [
                          "string",
                          "null"
                        ]
                      },
                      "claim": {
                        "anyOf": [
                          {
                            "type": "object",
                            "properties": {
                              "kind": {
                                "type": "string",
                                "enum": [
                                  "SELLER_PROCEEDS",
                                  "BUYER_REFUND"
                                ]
                              },
                              "status": {
                                "type": "string",
                                "enum": [
                                  "NONE",
                                  "CLAIMABLE",
                                  "WITHDRAWN"
                                ]
                              },
                              "beneficiary": {
                                "type": "string",
                                "pattern": "^0x[0-9a-fA-F]{40}$"
                              },
                              "amountWei": {
                                "type": "string",
                                "pattern": "^(0|[1-9]\\d*)$"
                              }
                            },
                            "required": [
                              "kind",
                              "status",
                              "beneficiary",
                              "amountWei"
                            ],
                            "additionalProperties": false
                          },
                          {
                            "type": "null"
                          }
                        ]
                      }
                    },
                    "required": [
                      "saleId",
                      "tokenId",
                      "seller",
                      "buyer",
                      "priceWei",
                      "fundedAt",
                      "expiresAt",
                      "status",
                      "tokenReclaimed",
                      "metadataStatus",
                      "catalogId",
                      "claim"
                    ],
                    "additionalProperties": false
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "freshness": {
                "type": "object",
                "properties": {
                  "projectionStatus": {
                    "type": "string",
                    "enum": [
                      "UNINITIALIZED",
                      "SYNCING",
                      "CURRENT",
                      "STALE",
                      "REBUILD_REQUIRED",
                      "REBUILDING",
                      "RECOVERY_REQUIRED"
                    ]
                  },
                  "observationFreshness": {
                    "type": "string",
                    "enum": [
                      "FRESH",
                      "STALE",
                      "UNKNOWN"
                    ]
                  },
                  "observationAgeSeconds": {
                    "anyOf": [
                      {
                        "type": "string",
                        "pattern": "^(0|[1-9]\\d*)$"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "lastObservedHead": {
                    "anyOf": [
                      {
                        "type": "string",
                        "pattern": "^(0|[1-9]\\d*)$"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "lagBlocks": {
                    "anyOf": [
                      {
                        "type": "string",
                        "pattern": "^(0|[1-9]\\d*)$"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "lastObservedAt": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "lastRpcSuccessAt": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "workerHeartbeatAt": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "recoveryReason": {
                    "type": [
                      "string",
                      "null"
                    ]
                  }
                },
                "required": [
                  "projectionStatus",
                  "observationFreshness",
                  "observationAgeSeconds",
                  "lastObservedHead",
                  "lagBlocks",
                  "lastObservedAt",
                  "lastRpcSuccessAt",
                  "workerHeartbeatAt",
                  "recoveryReason"
                ],
                "additionalProperties": false
              },
              "observation": {
                "type": "object",
                "properties": {
                  "requestEcho": {
                    "type": "object",
                    "properties": {
                      "transactionHash": {
                        "type": "string",
                        "pattern": "^0x[0-9a-fA-F]{64}$"
                      },
                      "blockNumber": {
                        "type": "string",
                        "pattern": "^(0|[1-9]\\d*)$"
                      },
                      "blockHash": {
                        "type": "string",
                        "pattern": "^0x[0-9a-fA-F]{64}$"
                      },
                      "logIndex": {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 9007199254740991
                      }
                    },
                    "required": [
                      "transactionHash",
                      "blockNumber",
                      "blockHash",
                      "logIndex"
                    ],
                    "additionalProperties": false
                  },
                  "coverage": {
                    "type": "string",
                    "enum": [
                      "NOT_REACHED",
                      "SCANNED",
                      "UNVERIFIABLE"
                    ]
                  },
                  "observedHeaderAtRequestedHeight": {
                    "anyOf": [
                      {
                        "type": "object",
                        "properties": {
                          "blockNumber": {
                            "type": "string",
                            "pattern": "^(0|[1-9]\\d*)$"
                          },
                          "blockHash": {
                            "type": "string",
                            "pattern": "^0x[0-9a-fA-F]{64}$"
                          },
                          "canonical": {
                            "type": "boolean"
                          },
                          "scanComplete": {
                            "type": "boolean"
                          }
                        },
                        "required": [
                          "blockNumber",
                          "blockHash",
                          "canonical",
                          "scanComplete"
                        ],
                        "additionalProperties": false
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "eventLookup": {
                    "type": "string",
                    "enum": [
                      "NOT_FOUND",
                      "MATCHED",
                      "NONCANONICAL",
                      "SELECTOR_MISMATCH"
                    ]
                  },
                  "matchedEvent": {
                    "anyOf": [
                      {
                        "type": "object",
                        "properties": {
                          "transactionHash": {
                            "type": "string",
                            "pattern": "^0x[0-9a-fA-F]{64}$"
                          },
                          "blockNumber": {
                            "type": "string",
                            "pattern": "^(0|[1-9]\\d*)$"
                          },
                          "blockHash": {
                            "type": "string",
                            "pattern": "^0x[0-9a-fA-F]{64}$"
                          },
                          "logIndex": {
                            "type": "integer",
                            "minimum": 0,
                            "maximum": 9007199254740991
                          },
                          "emitter": {
                            "type": "string",
                            "pattern": "^0x[0-9a-fA-F]{40}$"
                          },
                          "saleId": {
                            "type": "string",
                            "pattern": "^(0|[1-9]\\d*)$"
                          },
                          "buyer": {
                            "type": "string",
                            "pattern": "^0x[0-9a-fA-F]{40}$"
                          },
                          "amountWei": {
                            "type": "string",
                            "pattern": "^(0|[1-9]\\d*)$"
                          },
                          "fundedAt": {
                            "type": "string",
                            "pattern": "^(0|[1-9]\\d*)$"
                          },
                          "expiresAt": {
                            "type": "string",
                            "pattern": "^(0|[1-9]\\d*)$"
                          }
                        },
                        "required": [
                          "transactionHash",
                          "blockNumber",
                          "blockHash",
                          "logIndex",
                          "emitter",
                          "saleId",
                          "buyer",
                          "amountWei",
                          "fundedAt",
                          "expiresAt"
                        ],
                        "additionalProperties": false
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "projectionEffect": {
                    "type": "string",
                    "enum": [
                      "NOT_ASSESSED",
                      "CONSISTENT",
                      "INCONSISTENT"
                    ]
                  }
                },
                "required": [
                  "requestEcho",
                  "coverage",
                  "observedHeaderAtRequestedHeight",
                  "eventLookup",
                  "matchedEvent",
                  "projectionEffect"
                ],
                "additionalProperties": false
              }
            },
            "required": [
              "sale",
              "freshness",
              "observation"
            ],
            "additionalProperties": false
          },
          "provenance": {
            "type": "object",
            "properties": {
              "deploymentId": {
                "type": "string",
                "pattern": "^0x[0-9a-fA-F]{64}$"
              },
              "indexedBlockNumber": {
                "type": "string",
                "pattern": "^(0|[1-9]\\d*)$"
              },
              "indexedBlockHash": {
                "type": "string",
                "pattern": "^0x[0-9a-fA-F]{64}$"
              },
              "projectorVersion": {
                "type": "string",
                "pattern": "^(0|[1-9]\\d*)$"
              },
              "projectionBuildId": {
                "type": "string",
                "minLength": 1
              },
              "logScopeHash": {
                "type": "string",
                "pattern": "^0x[0-9a-fA-F]{64}$"
              }
            },
            "required": [
              "deploymentId",
              "indexedBlockNumber",
              "indexedBlockHash",
              "projectorVersion",
              "projectionBuildId",
              "logScopeHash"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "data",
          "provenance"
        ],
        "additionalProperties": false
      },
      "SalesEnvelope": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "saleId": {
                  "type": "string",
                  "pattern": "^(0|[1-9]\\d*)$"
                },
                "tokenId": {
                  "type": "string",
                  "pattern": "^(0|[1-9]\\d*)$"
                },
                "seller": {
                  "type": "string",
                  "pattern": "^0x[0-9a-fA-F]{40}$"
                },
                "buyer": {
                  "anyOf": [
                    {
                      "type": "string",
                      "pattern": "^0x[0-9a-fA-F]{40}$"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "priceWei": {
                  "type": "string",
                  "pattern": "^(0|[1-9]\\d*)$"
                },
                "fundedAt": {
                  "anyOf": [
                    {
                      "type": "string",
                      "pattern": "^(0|[1-9]\\d*)$"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "expiresAt": {
                  "anyOf": [
                    {
                      "type": "string",
                      "pattern": "^(0|[1-9]\\d*)$"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "LISTED",
                    "FUNDED",
                    "COMPLETED",
                    "CANCELLED",
                    "EXPIRED"
                  ]
                },
                "tokenReclaimed": {
                  "type": "boolean"
                },
                "metadataStatus": {
                  "type": "string",
                  "enum": [
                    "AVAILABLE",
                    "MISSING"
                  ]
                },
                "catalogId": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "claim": {
                  "anyOf": [
                    {
                      "type": "object",
                      "properties": {
                        "kind": {
                          "type": "string",
                          "enum": [
                            "SELLER_PROCEEDS",
                            "BUYER_REFUND"
                          ]
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "NONE",
                            "CLAIMABLE",
                            "WITHDRAWN"
                          ]
                        },
                        "beneficiary": {
                          "type": "string",
                          "pattern": "^0x[0-9a-fA-F]{40}$"
                        },
                        "amountWei": {
                          "type": "string",
                          "pattern": "^(0|[1-9]\\d*)$"
                        }
                      },
                      "required": [
                        "kind",
                        "status",
                        "beneficiary",
                        "amountWei"
                      ],
                      "additionalProperties": false
                    },
                    {
                      "type": "null"
                    }
                  ]
                }
              },
              "required": [
                "saleId",
                "tokenId",
                "seller",
                "buyer",
                "priceWei",
                "fundedAt",
                "expiresAt",
                "status",
                "tokenReclaimed",
                "metadataStatus",
                "catalogId",
                "claim"
              ],
              "additionalProperties": false
            }
          },
          "provenance": {
            "type": "object",
            "properties": {
              "deploymentId": {
                "type": "string",
                "pattern": "^0x[0-9a-fA-F]{64}$"
              },
              "indexedBlockNumber": {
                "type": "string",
                "pattern": "^(0|[1-9]\\d*)$"
              },
              "indexedBlockHash": {
                "type": "string",
                "pattern": "^0x[0-9a-fA-F]{64}$"
              },
              "projectorVersion": {
                "type": "string",
                "pattern": "^(0|[1-9]\\d*)$"
              },
              "projectionBuildId": {
                "type": "string",
                "minLength": 1
              },
              "logScopeHash": {
                "type": "string",
                "pattern": "^0x[0-9a-fA-F]{64}$"
              }
            },
            "required": [
              "deploymentId",
              "indexedBlockNumber",
              "indexedBlockHash",
              "projectorVersion",
              "projectionBuildId",
              "logScopeHash"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "data",
          "provenance"
        ],
        "additionalProperties": false
      },
...
      "ApiError": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message",
              "requestId"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      }
    }
  }
}